Symmetry hunt: corpus minor asymmetries 7 -> 1 - king proximity and bishop-pawns accumulators upgraded float->double (same class as the long-term mobility fix: square-order summation differs between mirrored boards, float epsilon flips the final int truncation). Last remaining case is a +-1cp pawn-structure TOTAL whose hidden float sum rounds differently from its displayed integer components - documented, telemetry keeps counting, tolerance covers it. Corpus: 2170 positions, 1 minor asymmetry

+++ opti_chess/eval_king.cpp
@@ -991,7 +991,10 @@ int Board::get_king_proximity()
 	update_kings_pos();
 
 	// King proximity
-	float proximity = 0.0f;
+	// double, not float: the per-pawn contributions are summed in SQUARE
+	// order, which differs between mirrored boards; float rounding noise
+	// (1e-7) used to flip the final int truncation by +-1cp on one side.
+	double proximity = 0.0;
 
 	// Proximity bonus for enemy pawns unprotected by another pawn
 	// TODO
